SETI-808: Ignore non-ASCII jobs in Jenkins
Python implementation of gearman server breaks down when trying to
register a non-ASCII job from workers. This is an upstream bug and
since we do not have our own fork of gear server, the easiest
solution is to ignore jobs with non-ASCII names.

These are already forbidden in zuul layout, which is the only
source jobs handed to gearman. Effectively, this is just a
security measure to prevent gearman server from failing in case
somebody accidentaly manually (not through gooddata/ci-infra repo)
adds a non-ASCII job to Jenkins.